Refactor wallet tab mod (gui)
